Develop

+++ app/src/js/app/shared/services/interceptor/interceptor.service.js
@@ -9,6 +9,10 @@
             response: function (response) {
                 // do something on success
                 var state = $injector.get('$state');
+                if (state && state.current && state.current.name == "app.tasks_trelloEdit") {
+                    response.isTrello = true;
+                }
+                    
                 $rootScope.$watch('state', function(newState, oldState){
                     if($window.localStorage["userId"] != $rootScope.userId ||
                     $window.localStorage["userName"] != $rootScope.userName || 
@@ -17,18 +21,20 @@
                     $window.localStorage["isAdmin"] != $rootScope.isAdmin ||
                     $window.localStorage["isClient"] != $rootScope.isClient ||
                     $window.localStorage["isDeveloper"] != $rootScope.isDeveloper){
-                        $window.localStorage.clear();
-                        state.go("login")
+                      //  $window.localStorage.clear();
+                      //  state.go("login")
                     }
                 }, true);
                 if(response.headers()['content-type'] == "application/json"){
                     if(response.status == 200){
-                        if (typeof response.data !== 'undefined') {
-                            if (response.data && response.data.response &&  !response.data.response.token && !response.data.response.email) {
+                        if (response && typeof response.data !== 'undefined') {
+                            if (response && response.data && response.data.response &&  !response.data.response.token && !response.data.response.email) {
+                                if (response.isTrello) {
+                                    return response;
+                                }else {
                                 var rest = $injector.get('UserServices');
                                 rest.persistence( function(error, response){
                                     var user = angular.copy(response);
-                                    console.log("SETEA LOCALSTORE::", user)
                                     $window.localStorage[TOKEN_KEY]   = user.token;
                                     $window.localStorage["userId"]    = user.id;
                                     $window.localStorage["userName"]  = user.name;
@@ -40,7 +46,8 @@
                                     $window.localStorage["idUserClient"]  =user.idClient;
                                     return;
                                 });
-                            } else if(response.data.code === 401 ||response.data.code === 403 ){
+                                }
+                            } else if(response && response.data && response.data.code === 401 ||response && response.data && response.data.code === 403 ){
                                 $window.localStorage.clear();
                                 state.go("login")                            
                             }
@@ -48,6 +55,7 @@
                     }                   
                 }
                 return response;
+                
             },
             responseError: function (response) {
                 // do something on error
